LayoutTests/platform/gtk/imported/w3c/web-platform-tests/css/cssom/getComputedStyle-detached-subtree-expected.txt


PASS getComputedStyle returns no style for detached element
FAIL getComputedStyle returns no style for element in non-rendered iframe (display: none) assert_equals: expected 0 but got 389
FAIL getComputedStyle returns no style for element in non-rendered iframe (display: none) from iframe's window assert_equals: expected 0 but got 389
FAIL getComputedStyle returns no style for element outside the flat tree assert_equals: expected 0 but got 389
FAIL getComputedStyle returns no style for descendant outside the flat tree assert_equals: expected 0 but got 389
PASS getComputedStyle returns no style for shadow tree outside of flattened tree
